Move dataflow assets and data into shared-assets folder.
