+++ src/plugins/dataflow/nodes/dataflow-node.scss
@@ -293,7 +293,7 @@ $node-inner-width: 140px;
     //  ON --- additional for when holding
     &.has-gate.gate-active .inputs {
       .number2 .socket {
-        background-image: url("../assets/icons/control/signal.svg");
+        background-image: url("../../shared-assets/icons/dataflow/control/signal.svg");
       }
       .binary .socket {
         transform:rotate(180deg);
@@ -304,7 +304,7 @@ $node-inner-width: 140px;
     }
 
     &.has-gate.gate-active .item.top .label {
-      background-image: url("../assets/icons/control/signal.svg");
+      background-image: url("../../shared-assets/icons/dataflow/control/signal.svg");
       background-repeat: no-repeat;
       background-position: 70px;
     }
