+++ src/plugins/simulator/simulations/brainwaves-gripper.scss
@@ -13,7 +13,7 @@
   }
 
   .gripper-image {
-    margin-left: -10px;
+    margin-left: -11px;
     margin-top: 20px;
   }
 }
